Feat/publish numbers

+++ internal/interface/dao/himo/user.go
@@ -2,6 +2,6 @@ package himo
 
 // User はユーザーの DAO
 type User struct {
-	ID       int64  `db:"id, primarykey, autoincrement"`
+	ID       int    `db:"id, primarykey, autoincrement"`
 	Nickname string `db:"nickname"`
 }
